Minor fix and a Feature addition
* Fix: Expand the file/post names.
If user has entered blog path as ~/path/to/blog when adding new post it
won't be getting added to the git because shell-command fails to expand
~. This was cross checked by giving full path to the blog and new posts
get added to Git. So if new post is not added to git *promote* post feature
does not work and we have to manually move posts from _drafts to _post. This
commit fixes it by using *expand-file-name*

* Proper post sorting by Jekyll
If we create to posts on same date Jekyll was sorting the posts based on
alphabets irrespective of when I created post. Basically jekyll was using
date from the post name. This bug on jekyll describes it [1]. To fix this
a *date* YAML front matter was introduced by jekyll.

This patch adds *date* YAML front matter with a proper date time string so
jekyll can consider this string and sort the pages properly on post timings

[1] https://github.com/mojombo/jekyll/issues/62