Update the date field on C-c C-c
Some times we create post and keep it in draft and want to publish it
later. But previous pull request I introduced date: header to order
the posts which are published on same day. Due to this if I publish
draft at some time later like says a week later it will be pushed down
in RSS feeds mainly sent for planet software. This is undesirable and
also this date field is meant to be the time stamp when post is
published so this commit will update the time stamp on hitting C-c C-c
in hyde-md.el